Move MainActivity to app module and expose UI screen
