+++ app/src/main/java/com/magter/myassistant/MainActivity.kt
@@ -1,9 +1,9 @@
-package com.magter.myassistant.ui
+package com.magter.myassistant
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material3.Text
+import com.magter.myassistant.ui.MainScreen
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -11,7 +11,7 @@ class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            Text("Hello from :feature:ui!")
+            MainScreen()
         }
     }
 }
