+++ feature/ui/src/main/java/com/magter/myassistant/ui/MainScreen.kt
@@ -1,0 +1,9 @@
+package com.magter.myassistant.ui
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MainScreen() {
+    Text("Hello from :feature:ui!")
+}
